Abstract class name with abstract prefix
